requirements.txt
pydantic-settings
python-telegram-bot
openai
loguru